Move menubutton to menu component
