+++ src/components/Menu/package.json
@@ -23,7 +23,7 @@
     },
     "dependencies": {
         "@gemeente-denhaag/baseprops": "^0.2.2",
-        "@material-ui/core": "^4.11.0"
+        "@gemeente-denhaag/icons": "^0.2.0"
     },
     "peerDependencies": {
         "react": "^17.0.1"
