+++ src/components/Menu/src/MenuButton.stories.mdx
@@ -1,10 +1,9 @@
-import { ArgsTable, Canvas, Description, Meta, Story } from "@storybook/addon-docs/blocks";
+import { ArgsTable, Canvas, Description, Meta, Story } from "@storybook/addon-docs";
 import { DesignTokenDocBlock } from "storybook-design-token/dist/doc-blocks";
 
-import MenuButton, { MenuButtonProps } from '.';
+import { MenuButton, MenuButtonProps } from '.';
 import pkg from '../package.json';
 import readme from "../README.md";
-
 
 <Meta
   title="Components/Navigation/MenuButton"
@@ -68,5 +67,28 @@ import readme from "../README.md";
       },
     }}>
       {(args) => <MenuButton { ... args} onclick={() => alert('Hello from MenuButton 👋')}>Label</MenuButton>}
+  </Story>
+</Canvas>
+
+### Menu
+<Canvas>
+  <Story name="Menu"
+    parameters={{
+      docs: {
+        source: {
+          code: `
+          `,
+
+         }
+      }
+    }}>
+      {(args)} => <Menu>
+                    <MenuButton>
+                      Test
+                    </MenuButton>
+                    <MenuButton>
+                      Test2
+                    </MenuButton>
+                  </Menu>
   </Story>
 </Canvas>
