+++ src/components/Menu/src/index.tsx
@@ -1,8 +1,9 @@
-import Menu from './Menu';
+import React from 'react';
+import BaseProps from '@gemeente-denhaag/baseprops';
+
+export const Menu: React.FC<BaseProps> = (props: BaseProps) => {
+  return <span>{props.children}</span>;
+};
 
 export default Menu;
-export * from './Menu';
-export * from './MenuItem';
-export * from './MenuList';
-export * from './Popover';
-export * from './Popper';
+export * from './MenuButton';
