--- src/components/MenuButton/package.json
@@ -1,32 +1,0 @@
-{
-    "name": "@gemeente-denhaag/menubutton",
-    "version": "0.1.0",
-    "description": "Buttons to be used in menu's",
-    "bugs": "https://github.com/Gemeente-DenHaag/denhaag-component-library/issues",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/Gemeente-DenHaag/denhaag-component-library.git",
-        "directory": "src/components/MenuButton"
-    },
-    "license": "EUPL-1.2",
-    "author": "Municipality of The Hague",
-    "exports": "./dist/index.js",
-    "main": "dist/index.js",
-    "module": "dist/index.js",
-    "types": "dist/index.d.ts",
-    "files": [
-        "dist"
-    ],
-    "scripts": {
-        "build": "rollup -c ../../../rollup.config.js",
-        "clean": "yarn rimraf dist tsconfig.tsbuildinfo"
-    },
-    "dependencies": {
-        "@gemeente-denhaag/icons": "^0.2.0"
-    },
-    "peerDependencies": {
-        "react": "^17.0.1"
-    },
-    "devDependencies": {
-    }
-}
