--- src/components/MenuButton/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "tsBuildInfoFile": "tsconfig.tsbuildinfo",
-    "composite": true
-  },
-  "include": ["src/**/*"],
-  "exclude": ["dist", "src/*.stories.tsx"],
-  "references": [
-    {
-      "path": "../Link"
-    }
-  ]
-}
